Move /systems folder beneath /lib
I moved the directory and fixed paths in code referencing the old
structure.

Paths got a little overcomplicated having /systems and /lib in the root
directory. The last straw came in loading 3rd-party scripts. NodeVM
requires an absolute path to the script in order for it to load modules
using a relative path. Inserting '..' to access the /systems directory
from /lib meant the path to the system's script became relative, which
broke loading modules by relative path. At this point, it became the
obvious choice to move the /systems directory beneath /lib.